Move the ExtendedKeyspaceMixin into a seperate contrib module and mark it as unsupported.
